+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 `.` minor | `-` Fix | `+` Addition | `^` improvement | `!` Change | `*` Refactor
 
 
+## 2024-06-10 - `0.0.14`
+
+- `-` ollama - improve Ollama Adapter to support multi system messages
+- `-` gemini - fix adapter to set "systemInstruction" (Supported in v1beta)
+
 ## 2024-06-10 - `0.0.13`
 
 - `+` Added AdapterKindResolver